Move helper methods into utils/helperMethods.js
- Rename addHistory to searchHistory
