+++ src/actions/ApiActions.js
@@ -1,7 +1,7 @@
 import { browserHistory } from 'react-router';
 import { SET_REF_RESULTS, SET_REF_HEADERS, SENDING_REF_REQUEST, SET_REF_QUERY, SET_REF_ERROR_MESSAGE } from '../constants/ApiConstants';
 import apiSearch from '../utils/apiSearch';
-import searchHistory from '../utils/addHistory';
+import searchHistory from '../utils/searchHistory';
 
 /**
  * Get info (version/last updated) from the Node server
