+++ src/utils/searchHistory.js
